requirements.txt
asyncio
python-dotenv
python-dotenv
openai
